refactor: 컨벤션에 맞춰 디렉토리 재정비
- conflict 해결하며 user 관련 패키징의 차이로 인한 오류 해결
