+++ app-main/src/main/java/net/causw/app/main/domain/user/account/service/dto/request/AdmissionResult.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.user.account.service.v2.dto;
+package net.causw.app.main.domain.user.account.service.dto.request;
 
 import java.time.LocalDateTime;
 import java.util.List;
