+++ app-main/src/main/java/net/causw/app/main/domain/user/account/service/dto/request/AdmissionStateResult.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.user.account.service.v2.dto;
+package net.causw.app.main.domain.user.account.service.dto.request;
 
 import net.causw.app.main.domain.user.account.entity.user.User;
 import net.causw.app.main.domain.user.account.enums.user.UserState;
